+++ WerewolvesCompany/Plugin.cs
@@ -76,22 +76,22 @@ namespace WerewolvesCompany
         private void ConfigSetup()
         {
             // Default parameters
-            config_DefaultInteractRange = Config.Bind("Default Interact Range", "Value", 1.0f, "How far the player can use his Action on another player.");
+            config_DefaultInteractRange = Config.Bind("Default Interact Range", "Value", 1.5f, "How far the player can use his Action on another player.");
             config_DefaultActionCoolDown = Config.Bind("Default Role Action Cooldown", "Value", 120f, "How often can a player use his action (in seconds).");
             config_DefaultStartOfRoundActionCoolDown = Config.Bind("Default Start of Round Role Action Cooldown", "Value", 120f, "How soon after the start of a round has started a player can use his action (in seconds).");
 
             // Werewolf parameters
-            config_WerewolfInteractRange = Config.Bind("Werewolf Kill Range", "Value", 1.0f, "How far a Werewolf can kill another player.");
+            config_WerewolfInteractRange = Config.Bind("Werewolf Kill Range", "Value", 1.5f, "How far a Werewolf can kill another player.");
             config_WerewolfActionCoolDown = Config.Bind("Werewolf Kill Cooldown", "Value", 120f, "How often a Werewolf can kill another player (in seconds).");
             config_WerewolfStartOfRoundActionCoolDown = Config.Bind("Werewolf Kill Cooldown at the start of a round", "Value", 120f, "How soon after the start of a round a Werewolf can Kill someone (in seconds).");
 
             // Villager parameters
-            config_VillagerInteractRange = Config.Bind("Villager Interact Range", "Value", 1.0f, "Unused -- How far the player can use his Action on another player");
+            config_VillagerInteractRange = Config.Bind("Villager Interact Range", "Value", 1.5f, "Unused -- How far the player can use his Action on another player");
             config_VillagerActionCoolDown = Config.Bind("Villager PatPat Cooldown", "Value", 0f, "Unused -- How often can a Villager use his action (in seconds).");
             config_VillagerStartOfRoundActionCoolDown = Config.Bind("Villager PatPat Cooldown at the start of a round", "Value", 0f, "How soon after a round has started a Villager can PatPat another villager (in seconds).");
 
             // Witch parameters
-            config_WitchInteractRange = Config.Bind("Witch Interact Range", "Value", 1.0f, "How far a Witch can use a potion on another.");
+            config_WitchInteractRange = Config.Bind("Witch Interact Range", "Value", 1.5f, "How far a Witch can use a potion on another.");
             config_WitchActionCoolDown = Config.Bind("Witch Potion Cooldown", "Value", 9999f, "How often a Witch can use each potion (in seconds).");
             config_WitchStartOfRoundActionCoolDown = Config.Bind("Witch Potion Cooldown at the start of a round", "Value", 120f, "How soon after a round has started a Witch can use her potions (in seconds).");