+++ WerewolvesCompany/Plugin.cs
@@ -23,7 +23,7 @@ namespace WerewolvesCompany
     {
         const string GUID = "doep.WerewolvesCompany";
         const string NAME = "WerewolvesCompany";
-        const string VERSION = "0.2.2";
+        const string VERSION = "0.2.3";
 
         internal static InputsClass InputActionsInstance;