+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### [0.2.3]
+
+
 ### [0.2.2]
 
 Custom death message when you die to a player's role.